requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Commented out - using SQLite for easier setup

# AI & ML
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
sentence-transformers==2.2.2
pinecone-client==2.2.4
weaviate-client==3.25.3

# Data Processing
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# APIs & Integrations
praw==7.7.1
PyGithub==1.59.1
sendgrid==6.10.0
twilio==8.10.0
discord.py==2.3.2

# Monitoring & Dashboard
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
schedule==1.2.0
loguru==0.7.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
